% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainy.R
\name{grain}
\alias{grain}
\title{Grain
Encapsulate a grid, with dimension and extent.}
\description{
A grid is a raster, a domain with an underlying regular discretization in xy.
}
\examples{

## ------------------------------------------------
## Method `grain$new`
## ------------------------------------------------

grain$new(c(360, 180), c(-180, 180, -90, 90))
 grain$new(c(360, 180), c(-180, 180, -90, 90))$align_extent(c(-179.5, 179, 0.03, 10))
g <- grain$new(world = TRUE)
g$cell_from_col(2)

## ------------------------------------------------
## Method `grain$cell_from_extent`
## ------------------------------------------------

g <- grain$new(world = TRUE)
g$cell_from_extent(c(147, 148, -43, -41))

## ------------------------------------------------
## Method `grain$xy_from_cell`
## ------------------------------------------------

g <- grain$new(world = TRUE)
g$xy_from_cell(g$cell_from_extent(c(147, 148, -43, -41)))
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dimension}}{dimension (nx, ny) of grid, number of columns, number of rows, defaults to 1x1 (unless 'world = TRUE')}

\item{\code{extent}}{extent of grid xmin,xmax,ymin,ymax - defaults to 0,ncol,0,nrow}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-grain-new}{\code{grain$new()}}
\item \href{#method-grain-align_extent}{\code{grain$align_extent()}}
\item \href{#method-grain-cell_from_col}{\code{grain$cell_from_col()}}
\item \href{#method-grain-cell_from_extent}{\code{grain$cell_from_extent()}}
\item \href{#method-grain-cell_from_row}{\code{grain$cell_from_row()}}
\item \href{#method-grain-cell_from_col_row}{\code{grain$cell_from_col_row()}}
\item \href{#method-grain-cell_from_colrow_combine}{\code{grain$cell_from_colrow_combine()}}
\item \href{#method-grain-cell_from_xy}{\code{grain$cell_from_xy()}}
\item \href{#method-grain-col_from_cell}{\code{grain$col_from_cell()}}
\item \href{#method-grain-col_from_x}{\code{grain$col_from_x()}}
\item \href{#method-grain-colrow_from_cell}{\code{grain$colrow_from_cell()}}
\item \href{#method-grain-extent_dimension}{\code{grain$extent_dimension()}}
\item \href{#method-grain-extent_from_cell}{\code{grain$extent_from_cell()}}
\item \href{#method-grain-extent_vrt}{\code{grain$extent_vrt()}}
\item \href{#method-grain-intersect_extent}{\code{grain$intersect_extent()}}
\item \href{#method-grain-listxyz}{\code{grain$listxyz()}}
\item \href{#method-grain-long}{\code{grain$long()}}
\item \href{#method-grain-n_cell}{\code{grain$n_cell()}}
\item \href{#method-grain-n_col}{\code{grain$n_col()}}
\item \href{#method-grain-n_row}{\code{grain$n_row()}}
\item \href{#method-grain-origin}{\code{grain$origin()}}
\item \href{#method-grain-plot_extent}{\code{grain$plot_extent()}}
\item \href{#method-grain-row_from_cell}{\code{grain$row_from_cell()}}
\item \href{#method-grain-row_from_y}{\code{grain$row_from_y()}}
\item \href{#method-grain-snap_extent}{\code{grain$snap_extent()}}
\item \href{#method-grain-vcrop}{\code{grain$vcrop()}}
\item \href{#method-grain-x_centre}{\code{grain$x_centre()}}
\item \href{#method-grain-x_corner}{\code{grain$x_corner()}}
\item \href{#method-grain-x_from_cell}{\code{grain$x_from_cell()}}
\item \href{#method-grain-x_from_col}{\code{grain$x_from_col()}}
\item \href{#method-grain-x_max}{\code{grain$x_max()}}
\item \href{#method-grain-x_min}{\code{grain$x_min()}}
\item \href{#method-grain-x_res}{\code{grain$x_res()}}
\item \href{#method-grain-xlim}{\code{grain$xlim()}}
\item \href{#method-grain-xy}{\code{grain$xy()}}
\item \href{#method-grain-xy_from_cell}{\code{grain$xy_from_cell()}}
\item \href{#method-grain-y_centre}{\code{grain$y_centre()}}
\item \href{#method-grain-y_corner}{\code{grain$y_corner()}}
\item \href{#method-grain-y_from_cell}{\code{grain$y_from_cell()}}
\item \href{#method-grain-y_from_row}{\code{grain$y_from_row()}}
\item \href{#method-grain-y_max}{\code{grain$y_max()}}
\item \href{#method-grain-y_min}{\code{grain$y_min()}}
\item \href{#method-grain-y_res}{\code{grain$y_res()}}
\item \href{#method-grain-ylim}{\code{grain$ylim()}}
\item \href{#method-grain-greet}{\code{grain$greet()}}
\item \href{#method-grain-clone}{\code{grain$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-new"></a>}}
\if{latex}{\out{\hypertarget{method-grain-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$new(dimension = NULL, extent = NULL, world = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dimension}}{ncol, nrow}

\item{\code{extent}}{xmin,xmax,ymin,ymax}

\item{\code{world}}{create a world longlat grid with 1 degree cells}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{grain$new(c(360, 180), c(-180, 180, -90, 90))
 grain$new(c(360, 180), c(-180, 180, -90, 90))$align_extent(c(-179.5, 179, 0.03, 10))
g <- grain$new(world = TRUE)
g$cell_from_col(2)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-align_extent"></a>}}
\if{latex}{\out{\hypertarget{method-grain-align_extent}{}}}
\subsection{Method \code{align_extent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$align_extent(x, snap = "out")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{x coordinate}

\item{\code{snap}}{"out", "in", "near"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
aligned extent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-cell_from_col"></a>}}
\if{latex}{\out{\hypertarget{method-grain-cell_from_col}{}}}
\subsection{Method \code{cell_from_col()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$cell_from_col(col)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{column index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
cell index, extent, or coordinate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-cell_from_extent"></a>}}
\if{latex}{\out{\hypertarget{method-grain-cell_from_extent}{}}}
\subsection{Method \code{cell_from_extent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$cell_from_extent(extent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{extent}}{xmin,xmax,ymin,ymax}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
cell index
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{g <- grain$new(world = TRUE)
g$cell_from_extent(c(147, 148, -43, -41))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-cell_from_row"></a>}}
\if{latex}{\out{\hypertarget{method-grain-cell_from_row}{}}}
\subsection{Method \code{cell_from_row()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$cell_from_row(row)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{row}}{row index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
cell index
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-cell_from_col_row"></a>}}
\if{latex}{\out{\hypertarget{method-grain-cell_from_col_row}{}}}
\subsection{Method \code{cell_from_col_row()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$cell_from_col_row(col, row)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{column index}

\item{\code{row}}{row index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
cell index
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-cell_from_colrow_combine"></a>}}
\if{latex}{\out{\hypertarget{method-grain-cell_from_colrow_combine}{}}}
\subsection{Method \code{cell_from_colrow_combine()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$cell_from_colrow_combine(col, row)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{column index}

\item{\code{row}}{row index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
cell index
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-cell_from_xy"></a>}}
\if{latex}{\out{\hypertarget{method-grain-cell_from_xy}{}}}
\subsection{Method \code{cell_from_xy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$cell_from_xy(xy)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xy}}{xy coordinates}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
cell index
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-col_from_cell"></a>}}
\if{latex}{\out{\hypertarget{method-grain-col_from_cell}{}}}
\subsection{Method \code{col_from_cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$col_from_cell(cell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell}}{cell index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
cell index
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-col_from_x"></a>}}
\if{latex}{\out{\hypertarget{method-grain-col_from_x}{}}}
\subsection{Method \code{col_from_x()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$col_from_x(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{x coordinate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
col index
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-colrow_from_cell"></a>}}
\if{latex}{\out{\hypertarget{method-grain-colrow_from_cell}{}}}
\subsection{Method \code{colrow_from_cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$colrow_from_cell(cell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell}}{cell index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
col,row index
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-extent_dimension"></a>}}
\if{latex}{\out{\hypertarget{method-grain-extent_dimension}{}}}
\subsection{Method \code{extent_dimension()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$extent_dimension(extent, snap = "out")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{extent}}{xmin,xmax,ymin,ymax}

\item{\code{snap}}{"out", "in", "near"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dimension
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-extent_from_cell"></a>}}
\if{latex}{\out{\hypertarget{method-grain-extent_from_cell}{}}}
\subsection{Method \code{extent_from_cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$extent_from_cell(cell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell}}{cell index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
extent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-extent_vrt"></a>}}
\if{latex}{\out{\hypertarget{method-grain-extent_vrt}{}}}
\subsection{Method \code{extent_vrt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$extent_vrt(vrt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vrt}}{vrt file path}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
extents, in a matrix, xmin,xmax,ymin,ymax
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-intersect_extent"></a>}}
\if{latex}{\out{\hypertarget{method-grain-intersect_extent}{}}}
\subsection{Method \code{intersect_extent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$intersect_extent(extent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{extent}}{xmin,xmax,ymin,ymax}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
extent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-listxyz"></a>}}
\if{latex}{\out{\hypertarget{method-grain-listxyz}{}}}
\subsection{Method \code{listxyz()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$listxyz(data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{optional data values for output}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of x,y vectors and z matrix (like \code{\link[graphics:image]{graphics::image()}})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-long"></a>}}
\if{latex}{\out{\hypertarget{method-grain-long}{}}}
\subsection{Method \code{long()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$long(data = NULL, raster_order = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{optional data values for output}

\item{\code{raster_order}}{raster order or traditional R matrix order}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data frame of x, y and optionaly z value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-n_cell"></a>}}
\if{latex}{\out{\hypertarget{method-grain-n_cell}{}}}
\subsection{Method \code{n_cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$n_cell()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer, number of cells
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-n_col"></a>}}
\if{latex}{\out{\hypertarget{method-grain-n_col}{}}}
\subsection{Method \code{n_col()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$n_col()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer, number of columns
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-n_row"></a>}}
\if{latex}{\out{\hypertarget{method-grain-n_row}{}}}
\subsection{Method \code{n_row()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$n_row()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer, number of rows
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-origin"></a>}}
\if{latex}{\out{\hypertarget{method-grain-origin}{}}}
\subsection{Method \code{origin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$origin()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
coordinate of grid origin
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-plot_extent"></a>}}
\if{latex}{\out{\hypertarget{method-grain-plot_extent}{}}}
\subsection{Method \code{plot_extent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$plot_extent(..., extent = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{args passed to \code{\link[graphics:plot.default]{graphics::plot()}}}

\item{\code{extent}}{xmin,xmax,ymin,ymax}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing, called for side effect of plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-row_from_cell"></a>}}
\if{latex}{\out{\hypertarget{method-grain-row_from_cell}{}}}
\subsection{Method \code{row_from_cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$row_from_cell(cell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell}}{cell index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
integer of row index
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-row_from_y"></a>}}
\if{latex}{\out{\hypertarget{method-grain-row_from_y}{}}}
\subsection{Method \code{row_from_y()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$row_from_y(y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{y coordinate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
integer of row index
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-snap_extent"></a>}}
\if{latex}{\out{\hypertarget{method-grain-snap_extent}{}}}
\subsection{Method \code{snap_extent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$snap_extent(res)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{res}}{resolution}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
extent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-vcrop"></a>}}
\if{latex}{\out{\hypertarget{method-grain-vcrop}{}}}
\subsection{Method \code{vcrop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$vcrop(extent, snap = "out")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{extent}}{xmin,xmax,ymin,ymax}

\item{\code{snap}}{"out", "in", "near"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
extent, dimension cropped (and snapped to grain)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-x_centre"></a>}}
\if{latex}{\out{\hypertarget{method-grain-x_centre}{}}}
\subsection{Method \code{x_centre()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$x_centre()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
x coordinates of centre of cell
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-x_corner"></a>}}
\if{latex}{\out{\hypertarget{method-grain-x_corner}{}}}
\subsection{Method \code{x_corner()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$x_corner()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
x coordinates of corners (edges) of cell
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-x_from_cell"></a>}}
\if{latex}{\out{\hypertarget{method-grain-x_from_cell}{}}}
\subsection{Method \code{x_from_cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$x_from_cell(cell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell}}{cell index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
x coordinate of cell (centre)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-x_from_col"></a>}}
\if{latex}{\out{\hypertarget{method-grain-x_from_col}{}}}
\subsection{Method \code{x_from_col()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$x_from_col(col)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{column index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
x of column (centre)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-x_max"></a>}}
\if{latex}{\out{\hypertarget{method-grain-x_max}{}}}
\subsection{Method \code{x_max()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$x_max()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
maximum x (right of grid)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-x_min"></a>}}
\if{latex}{\out{\hypertarget{method-grain-x_min}{}}}
\subsection{Method \code{x_min()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$x_min()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
minimum x (left of grid)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-x_res"></a>}}
\if{latex}{\out{\hypertarget{method-grain-x_res}{}}}
\subsection{Method \code{x_res()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$x_res()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
x resolution (width of cell)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-xlim"></a>}}
\if{latex}{\out{\hypertarget{method-grain-xlim}{}}}
\subsection{Method \code{xlim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$xlim()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
x range of grid (leftmost edge to rightmost edge)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-xy"></a>}}
\if{latex}{\out{\hypertarget{method-grain-xy}{}}}
\subsection{Method \code{xy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$xy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
coordinates of cells (centre)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-xy_from_cell"></a>}}
\if{latex}{\out{\hypertarget{method-grain-xy_from_cell}{}}}
\subsection{Method \code{xy_from_cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$xy_from_cell(cell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell}}{cell index}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{g <- grain$new(world = TRUE)
g$xy_from_cell(g$cell_from_extent(c(147, 148, -43, -41)))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-y_centre"></a>}}
\if{latex}{\out{\hypertarget{method-grain-y_centre}{}}}
\subsection{Method \code{y_centre()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$y_centre()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
y coordinate (centre)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-y_corner"></a>}}
\if{latex}{\out{\hypertarget{method-grain-y_corner}{}}}
\subsection{Method \code{y_corner()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$y_corner()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
y coordinates (corners)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-y_from_cell"></a>}}
\if{latex}{\out{\hypertarget{method-grain-y_from_cell}{}}}
\subsection{Method \code{y_from_cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$y_from_cell(cell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell}}{cell index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
y coordinate of cell (centre)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-y_from_row"></a>}}
\if{latex}{\out{\hypertarget{method-grain-y_from_row}{}}}
\subsection{Method \code{y_from_row()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$y_from_row(row)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{row}}{row index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
y coordinate of row (centre)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-y_max"></a>}}
\if{latex}{\out{\hypertarget{method-grain-y_max}{}}}
\subsection{Method \code{y_max()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$y_max()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
maximum y coordinate of grid (top edge)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-y_min"></a>}}
\if{latex}{\out{\hypertarget{method-grain-y_min}{}}}
\subsection{Method \code{y_min()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$y_min()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
minimum y coordinate of grid (left edge)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-y_res"></a>}}
\if{latex}{\out{\hypertarget{method-grain-y_res}{}}}
\subsection{Method \code{y_res()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$y_res()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
resolution y of grid (height of cell)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-ylim"></a>}}
\if{latex}{\out{\hypertarget{method-grain-ylim}{}}}
\subsection{Method \code{ylim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$ylim()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
extent of grid in y (bottom most edge to topmost edge)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-greet"></a>}}
\if{latex}{\out{\hypertarget{method-grain-greet}{}}}
\subsection{Method \code{greet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$greet()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
hello
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grain-clone"></a>}}
\if{latex}{\out{\hypertarget{method-grain-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
