% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainy-package.R
\docType{package}
\name{grainy-package}
\alias{grainy}
\alias{grainy-package}
\title{grainy: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/grainy}
  \item Report bugs at \url{https://github.com/mdsumner/grainy/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
