% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainy-package.R
\docType{package}
\name{grainy-package}
\alias{grainy}
\alias{grainy-package}
\title{grainy: Raster Stuff}
\description{
Encapsulate raster grid logic in an object, an object comprised of dimension and extent: a discretized domain. Create coordinates, query cell index, snap or align extents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/grainy}
  \item Report bugs at \url{https://github.com/mdsumner/grainy/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
